# docs/openapi.yml
openapi: 3.0.3
info:
  title: Super Notes API
  version: 1.0.0
  description: 'Notes that sync across devices, and answers to questions about them
    with citations. Sign in with `POST /api/v1/auth/google/` and send `Authorization:
    Bearer <access>`.'
paths:
...
  /api/v1/me/:
    get:
      operationId: me_retrieve
      description: The signed-in account.
      summary: Your profile
      tags:
      - Account
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: Not signed in.
    patch:
      operationId: me_partial_update
      description: '`timezone` (an IANA name) and `memory_enabled`. Setting `memory_enabled`
        also records that the user chose it themselves. Switching memory off stops
        it being learned and used; the facts stay until deleted (`DELETE memory/facts/`).'
      summary: Update your settings
      tags:
      - Account
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedMeUpdateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedMeUpdateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedMeUpdateRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
          description: ''
        '400':
          description: Invalid timezone.
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: Not signed in.
...
  /api/v1/notes/{id}/attachments/:
    get:
      operationId: notes_attachments_list
      description: Newest first. Metadata only; download a file with `GET attachments/<id>/file/`.
      summary: List a note's attachments
      parameters:
      - name: cursor
        required: false
        in: query
        description: The pagination cursor value.
        schema:
          type: string
      - in: path
        name: id
        schema:
          type: integer
        required: true
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      tags:
      - Attachments
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAttachmentList'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: No such note of yours.
    post:
      operationId: notes_attachments_create
      description: |-
        One file in the `file` field of a multipart form: a JPEG, PNG or WebP image, or a PDF, of at most 10 MB. The type is read from the file's bytes; the name and Content-Type sent are not trusted, and a file of any other type is refused with 415 whatever it is called. The name is kept for display and downloads, cleaned, and always ends in the type's extension.

        The same file again on the same note is not stored twice: you get the existing attachment with 200. A new one is 201, `status: pending`. Its text is then read in the background (`extracting`), and once it is `ready` search and Ask find it, naming the file; a file whose text cannot be read is `failed`, with the reason in `error`, and stays attached. It counts against your `storage_bytes` (see `me/`) until it is deleted. The note is sent again by `notes/changes/` with its attachments, on every status change too.
      summary: Attach a file to a note
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - Attachments
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AttachmentUploadRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attachment'
          description: Stored.
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attachment'
          description: This file is already on the note.
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: No file, or an empty one.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: No such note of yours.
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: '`too_large`: over the cap.'
        '415':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: '`unsupported_file_type`: not a JPEG, PNG, WebP or PDF; or
            `unsupported_media_type`: the request is not multipart.'
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorageQuotaExceeded'
          description: '`quota_exceeded`: your storage is full (with usage, in bytes),
            or `throttled`.'
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: '`system_limit_reached`: this feature is paused for everyone
            (a system-wide limit). Not the caller''s own quota; try again later.'
...
  /api/v1/notes/changes/:
    get:
      operationId: notes_changes_retrieve
      description: 'Every note written after revision `after`, deleted ones included
        as tombstones (no title or content, `deleted_at` set), oldest write first.
        A live note carries all its `reminders` and `attachments` (metadata only);
        adding, changing, finishing or deleting either sends its note again, so replace
        the note''s reminders and attachments with the ones sent. Store `latest_revision`
        and send it as `after` next time. If `has_more` is true, call again straight
        away with it: the batch was cut at `limit`. A first sync sends `after=0`.'
      summary: What changed since a revision
      parameters:
      - in: query
        name: after
        schema:
          type: integer
        description: 'Default 0: everything.'
      - in: query
        name: limit
        schema:
          type: integer
        description: 1-1000, default 500.
      tags:
      - Notes
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteChanges'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: ''
...
components:
  schemas:
...
    AskUsage:
      type: object
      description: 'This month''s asks: the `chat_turns` limit. Failed asks are not
        counted.'
      properties:
        used:
          type: integer
        limit:
          type: integer
          nullable: true
          description: 'Null: unlimited.'
        resets_at:
          type: string
          format: date-time
          nullable: true
          description: 'When `used` goes back to 0: the start of next month, Asia/Kolkata.
            Null if it never resets.'
      required:
      - limit
      - resets_at
      - used
    Attachment:
      type: object
      description: |-
        An attachment's metadata, as every endpoint (and ``changes``) returns it.

        Never its bytes (``GET attachments/<id>/file/``) and never its extracted
        text: that is for search, and can be long.
      properties:
        id:
          type: integer
          readOnly: true
        note:
          type: integer
          readOnly: true
        original_name:
          type: string
          readOnly: true
          description: The uploaded file's name, cleaned; always ends in the type's
            extension.
        mime_type:
          type: string
          readOnly: true
          description: Sniffed from the file's bytes, not taken from the upload.
        size:
          type: integer
          readOnly: true
          description: Bytes.
        sha256:
          type: string
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/AttachmentStatusEnum'
          readOnly: true
        error:
          type: string
          readOnly: true
        summary:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - error
      - id
      - mime_type
      - note
      - original_name
      - sha256
      - size
      - status
      - summary
    AttachmentStatusEnum:
      enum:
      - pending
      - extracting
      - ready
      - failed
      type: string
      description: |-
        * `pending` - Pending
        * `extracting` - Extracting
        * `ready` - Ready
        * `failed` - Failed
    AttachmentUploadRequest:
      type: object
      description: '``POST notes/<id>/attachments/``: one file, as multipart form
        data.'
      properties:
        file:
          type: string
          format: binary
          description: A JPEG, PNG, WebP image or a PDF. The type is read from the
            bytes; the name and Content-Type you send are not trusted.
      required:
      - file
...
    LimitUsage:
      type: object
      properties:
        used:
          type: integer
          description: Used this period (bytes for `storage_bytes`).
        limit:
          type: integer
          nullable: true
          description: 'Null: unlimited.'
        resets_at:
          type: string
          format: date-time
          nullable: true
          description: When `used` goes back to 0. Null if it never resets.
      required:
      - limit
      - resets_at
      - used
    Me:
      type: object
      description: The signed-in account, as a client shows it.
      properties:
        id:
          type: integer
          readOnly: true
        email:
          type: string
          format: email
          readOnly: true
        name:
          type: string
          readOnly: true
        avatar_url:
          type: string
          format: uri
          readOnly: true
        plan:
          allOf:
          - $ref: '#/components/schemas/PlanEnum'
          readOnly: true
        date_joined:
          type: string
          format: date-time
          readOnly: true
        timezone:
          type: string
          readOnly: true
        memory_enabled:
          type: boolean
          readOnly: true
        memory_choice_explicit:
          type: boolean
          readOnly: true
        ask_usage:
          allOf:
          - $ref: '#/components/schemas/AskUsage'
          readOnly: true
          description: 'This month''s asks: the same numbers as `limits.chat_turns`
            (assistant/quota.py), kept for V1 clients.'
        limits:
          allOf:
          - $ref: '#/components/schemas/UserLimits'
          readOnly: true
          description: Use and limit of each user-facing limit key (DECISIONS D84).
      required:
      - ask_usage
      - avatar_url
      - date_joined
      - email
      - id
      - limits
      - memory_choice_explicit
      - memory_enabled
      - name
      - plan
      - timezone
    Message:
      type: object
      description: 'A plain answer: what happened, and a stable code a client can
        branch on.'
      properties:
        detail:
          type: string
        code:
          type: string
      required:
      - code
      - detail
...
    NoteChange:
      oneOf:
      - $ref: '#/components/schemas/NoteSync'
      - $ref: '#/components/schemas/NoteTombstone'
    NoteChanges:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/NoteChange'
        latest_revision:
          type: integer
        has_more:
          type: boolean
      required:
      - has_more
      - latest_revision
      - results
...
    NoteSync:
      type: object
      description: A live note in ``changes``, with its reminders and attachments
        (deleted ones left out).
      properties:
        id:
          type: integer
          readOnly: true
        type:
          $ref: '#/components/schemas/TypeEnum'
        title:
          type: string
          maxLength: 500
        content: {}
        content_text:
          type: string
          readOnly: true
        version:
          type: integer
          readOnly: true
        revision:
          type: integer
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        deleted_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        reminders:
          type: array
          items:
            $ref: '#/components/schemas/Reminder'
          readOnly: true
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
          readOnly: true
      required:
      - attachments
      - content_text
      - created_at
      - deleted_at
      - id
      - reminders
      - revision
      - updated_at
      - version
    NoteTombstone:
      type: object
      description: 'A deleted note in ``changes``: enough to drop it, none of its
        content.'
      properties:
        id:
          type: integer
          readOnly: true
        type:
          allOf:
          - $ref: '#/components/schemas/TypeEnum'
          readOnly: true
        version:
          type: integer
          readOnly: true
        revision:
          type: integer
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        deleted_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - deleted_at
      - id
      - revision
      - type
      - updated_at
      - version
...
    PaginatedAttachmentList:
      type: object
      required:
      - results
      properties:
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?cursor=cD00ODY%3D"
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?cursor=cj0xJnA9NDg3
        results:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
...
    PatchedMeUpdateRequest:
      type: object
      description: What ``PATCH me/`` accepts. Anything else is ignored.
      properties:
        timezone:
          type: string
          minLength: 1
          description: An IANA timezone name, e.g. `Asia/Kolkata`.
          maxLength: 64
        memory_enabled:
          type: boolean
          description: Also marks the choice as the user's own (`memory_choice_explicit`).
...
    PlanEnum:
      enum:
      - free
      - premium
      type: string
      description: |-
        * `free` - Free
        * `premium` - Premium
...
    Reminder:
      type: object
      description: |-
        A reminder, as every endpoint returns it.

        Keep this the only schema component with a reminder ``status``: the
        schema's enums are named by field, and a second component carrying it
        would collide with the asks' ``status`` (see ReminderInRangeSerializer).
      properties:
        id:
          type: integer
          readOnly: true
        note:
          type: integer
          readOnly: true
        due_at:
          type: string
          format: date-time
          readOnly: true
        lead_days:
          type: integer
          readOnly: true
        channels:
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/ReminderStatusEnum'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - channels
      - created_at
      - due_at
      - id
      - lead_days
      - note
      - status
      - updated_at
...
    ReminderStatusEnum:
      enum:
      - scheduled
      - done
      - cancelled
      type: string
      description: |-
        * `scheduled` - Scheduled
        * `done` - Done
        * `cancelled` - Cancelled
...
    StorageQuotaExceeded:
      type: object
      description: 'A plain answer: what happened, and a stable code a client can
        branch on.'
      properties:
        detail:
          type: string
        code:
          type: string
        used:
          type: integer
          description: Used this period (bytes for `storage_bytes`).
        limit:
          type: integer
          nullable: true
          description: 'Null: unlimited.'
        resets_at:
          type: string
          format: date-time
          nullable: true
          description: When `used` goes back to 0. Null if it never resets.
      required:
      - code
      - detail
      - limit
      - resets_at
      - used
...
    TypeEnum:
      enum:
      - text
      - checklist
      type: string
      description: |-
        * `text` - Text
        * `checklist` - Checklist
...
    UserLimits:
      type: object
      description: One entry per user-facing limit key (DECISIONS D84, D91).
      properties:
        chat_turns:
          allOf:
          - $ref: '#/components/schemas/LimitUsage'
          description: Asks and chat turns, per month.
        format:
          allOf:
          - $ref: '#/components/schemas/LimitUsage'
          description: Format my note, per month.
        summary:
          allOf:
          - $ref: '#/components/schemas/LimitUsage'
          description: Attachment summaries, per month.
        storage_bytes:
          allOf:
          - $ref: '#/components/schemas/LimitUsage'
          description: Attachment storage in bytes, all time.
      required:
      - chat_turns
      - format
      - storage_bytes
      - summary
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT